gpio: litex: Make the irqchip immutable
Commit 6c846d026d49 ("gpio: Don't fiddle with irqchips marked as
immutable") added a warning to indicate if the gpiolib is altering the
internals of irqchips.  Following this change the following warning is
now observed for the litex driver:

    gpio gpiochip1: (litex_gpio): not an immutable chip, please consider fixing it!

Fix this by making the irqchip in the litex driver immutable.

While at it, use the driver's name instead of the platform device's
name, as that is the recommended name, and there is no
backwards-compatibility to take into account.

Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>